requirements.txt
streamlit==1.37.0
requests==2.32.3
aiohttp==3.9.5
numpy==1.26.4
google-generativeai==0.7.2
python-dotenv==1.0.1
beautifulsoup4==4.12.3
sentence-transformers==3.0.1
scikit-learn==1.5.1
selenium==4.23.1
langchain==0.2.10
langchain-google-genai==1.0.7
langchain-core==0.2.22
langchain-community==0.2.9  
langchain-text-splitters==0.2.2
streamlit-extras==0.4.3
python-dotenv
PyPDF2
text
faiss-cpu
